Extract RoborazziOptions for readability
